hw_collection - task 1
